WIP: Rework the whole project architecture
