+++ .gitignore
@@ -266,3 +266,4 @@ temp/
 demos/Thoth\.Elmish\.Demo/output/
 
 \.ionide/
+fableBuild/
