+++ Directory.Build.props
@@ -1,0 +1,27 @@
+<Project>
+	<PropertyGroup>
+		<!-- Make all the projects use a lock file -->
+		<RestorePackagesWithLockFile>true</RestorePackagesWithLockFile>
+		<!-- Make all the projects use the highest warning level -->
+		<WarningLevel>5</WarningLevel>
+	</PropertyGroup>
+    <!-- Packages common to all the projects  -->
+    <ItemGroup>
+        <!-- Reproducible builds + SourceLink supports -->
+        <!-- <PackageReference Include="DotNet.ReproducibleBuilds" Version="0.1.26" PrivateAssets="All"/> -->
+        <!-- Depends on the same version of FSharp.Core for all the projects -->
+        <!-- We don't use any feature related to FSharp.Core 5 so we force to use a lower one -->
+        <PackageReference Update="FSharp.Core" Version="4.7.2" />
+    </ItemGroup>
+	<!-- NuGet Metadata -->
+	<PropertyGroup>
+        <PackageProjectUrl>https://mangelmaxime.github.io/Nacara/</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/MangelMaxime/Nacara</RepositoryUrl>
+        <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
+		<GenerateDocumentationFile>true</GenerateDocumentationFile>
+		<Authors>Maxime Mangel</Authors>
+	</PropertyGroup>
+	<ItemGroup>
+		<None Include="$(MSBuildThisFileDirectory)\LICENSE.txt" Pack="true" PackagePath="$(PackageLicenseFile)" />
+	</ItemGroup>
+</Project>
