+++ package.json
@@ -1,53 +1,45 @@
 {
-  "name": "nacara",
-  "version": "0.4.1",
-  "main": "dist/Main.js",
-  "license": "MIT",
-  "bin": {
-    "nacara": "./cli.js"
-  },
-  "scripts": {
-    "nacara": "fable-splitter -c src/Nacara/splitter.config.js",
-    "nacara-watch": "fable-splitter -c src/Nacara/splitter.config.js --watch",
-    "layouts-standard": "fable-splitter -c src/Layouts/Standard/splitter.config.js",
-    "layouts-standard-watch": "fable-splitter -c src/Layouts/Standard/splitter.config.js --watch",
-    "nodemon": "npx nodemon cli.js",
-    "build": "yarn nacara && yarn layouts-standard"
-  },
-  "devDependencies": {
-    "@babel/plugin-transform-modules-commonjs": "^7.5.0",
-    "@types/react": "^16.9.2",
-    "@types/react-dom": "^16.9.0",
-    "fable-compiler": "^2.3.23",
-    "fable-splitter": "^2.1.11",
-    "nodemon": "^1.19.2",
-    "npx": "^10.2.0",
-    "tslint": "^5.19.0",
-    "tslint-react": "^4.0.0",
-    "typescript": "^3.6.2"
-  },
-  "dependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
-    "@babel/plugin-transform-typescript": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.5.5",
-    "@fortawesome/fontawesome-free": "^5.10.2",
-    "bulma": "^0.7.5",
-    "bulma-switch": "^2.0.0",
-    "chalk": "^2.4.2",
-    "chokidar": "^3.0.2",
-    "code-lightner": "^0.3.0",
-    "core-js": "^3.2.1",
-    "front-matter": "^3.0.2",
-    "live-server": "^1.2.1",
-    "markdown-it": "^9.1.0",
-    "markdown-it-container": "^2.0.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "sass": "^1.23.0-module.beta.1",
-    "slugify": "^1.3.4"
-  }
+    "name": "nacara-root",
+    "private": "true",
+    "license": "Apache-2.0",
+    "scripts": {
+        "postinstall": "dotnet tool restore",
+        "clean": "shx rm -rf dist",
+        "nacara": "fable-splitter -c src/Nacara/splitter.config.js",
+        "nacara-watch": "fable-splitter -c src/Nacara/splitter.config.js --watch",
+        "layouts-standard": "fable-splitter -c src/Layouts/Standard/splitter.config.js",
+        "layouts-standard-watch": "fable-splitter -c src/Layouts/Standard/splitter.config.js --watch",
+        "watch-nacara-layout": "dotnet fable src/Layouts/Standard --outDir src/Layouts/Standard/fableBuild --watch",
+        "watch-nacara-fable": "dotnet fable src/Nacara/Source --outDir src/Nacara/fableBuild --watch",
+        "watch-nacara-daemon": "npx nodemon --watch src/Nacara/fableBuild --watch src/Layouts/Standard/fableBuild --watch nacara.config.js --delay 150ms --exec \"npx nacara\"",
+        "watch": "concurrently -p none 'npm:watch-*'",
+        "build": "yarn nacara && yarn layouts-standard"
+    },
+    "devDependencies": {
+        "concurrently": "^6.1.0",
+        "nodemon": "^2.0.7",
+        "shx": "^0.3.3"
+    },
+    "engines": {
+        "node": ">=14.0.0",
+        "npm": ">=7.13.0"
+    },
+    "dependencies": {
+        "@fortawesome/fontawesome-free": "^5.15.3",
+        "bulma": "^0.9.2",
+        "bulma-switch": "^2.0.0",
+        "chalk": "^4.1.1",
+        "chokidar": "^3.5.1",
+        "code-lightner": "^0.3.0",
+        "core-js": "^3.12.1",
+        "esm": "^3.2.25",
+        "front-matter": "^4.0.2",
+        "live-server": "^1.2.1",
+        "markdown-it": "^12.0.6",
+        "markdown-it-container": "^3.0.0",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
+        "sass": "^1.32.13",
+        "slugify": "^1.5.3"
+    }
 }
