+++ src/Layouts/Standard/Export.fs
@@ -16,6 +16,6 @@ type IExport =
 exportDefault
     {
         new IExport with
-            member __.Default (model, pageContext) = Default.toHtml model pageContext
-            member __.Changelog (model, pageContext) = Changelog.toHtml model pageContext
+            member __.Default (model, pageContext) = Layout.Standard.Default.toHtml model pageContext
+            member __.Changelog (model, pageContext) = Layout.Standard.Changelog.toHtml model pageContext
     }
