+++ src/Layouts/Standard/package.json
@@ -1,0 +1,16 @@
+{
+    "name": "nacara-layout-standard",
+    "version": "0.0.0",
+    "description": "",
+    "main": "fableBuild/Export.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "Maxime Mangel <mangel.maxime@protonmail.com>",
+    "license": "Apache-2.0",
+    "peerDependencies": {
+        "bulma": "^0.9.2",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2"
+    }
+}
