--- src/Layouts/Standard/paket.references
@@ -1,10 +1,0 @@
-FSharp.Core
-Fable.Core
-Fable.Date
-Fable.React
-Fable.Node
-Fulma
-Fable.FontAwesome
-Fable.FontAwesome.Free
-Fable.Promise
-Thoth.Json
