--- src/Layouts/Standard/splitter.config.js
@@ -1,9 +1,0 @@
-const path = require("path");
-
-module.exports = {
-    entry: path.join(__dirname, "./Nacara.Layouts.Standard.fsproj"),
-    outDir: path.join(__dirname, "./../../../dist/layouts/standard"),
-    babel: {
-        plugins: ["@babel/plugin-transform-modules-commonjs"],
-    }
-};
