+++ src/Nacara/Source/Docs.fsproj
@@ -18,5 +18,16 @@
     <Compile Include="Write.fs" />
     <Compile Include="Main.fs" />
   </ItemGroup>
-  <Import Project="..\..\.paket\Paket.Restore.targets" />
+  <ItemGroup>
+    <PackageReference Include="Fable.Core" Version="3.2.7" />
+    <PackageReference Include="Fable.Date" Version="1.0.0" />
+    <PackageReference Include="Fable.Elmish" Version="3.1.0" />
+    <PackageReference Include="Fable.FontAwesome" Version="2.0.0" />
+    <PackageReference Include="Fable.FontAwesome.Free" Version="2.1.0" />
+    <PackageReference Include="Fable.Node" Version="1.2.0" />
+    <PackageReference Include="Fable.Promise" Version="2.2.0" />
+    <PackageReference Include="Fable.React" Version="7.4.0" />
+    <PackageReference Include="Fulma" Version="2.10.1" />
+    <PackageReference Include="Thoth.Json" Version="5.1.0" />
+  </ItemGroup>
 </Project>
