+++ src/Nacara/package.json
@@ -1,0 +1,28 @@
+{
+    "name": "nacara",
+    "version": "0.4.1",
+    "main": "fableBuild/Main.js",
+    "license": "MIT",
+    "bin": {
+        "nacara": "./cli.js"
+    },
+    "scripts": {},
+    "devDependencies": {},
+    "engines": {
+        "node": ">=14.0.0",
+        "npm": ">=7.13.0"
+    },
+    "dependencies": {
+        "chalk": "^4.1.1",
+        "chokidar": "^3.5.1",
+        "code-lightner": "^0.3.0",
+        "core-js": "^3.13.1",
+        "esm": "^3.2.25",
+        "front-matter": "^4.0.2",
+        "live-server": "^1.2.1",
+        "markdown-it": "^12.0.6",
+        "markdown-it-container": "^3.0.0",
+        "sass": "^1.34.0",
+        "slugify": "^1.5.3"
+    }
+}
